backend/requirements.txt
# Be Star Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.3
weasyprint==60.2
qrcode[pil]==7.4.2
aiosmtplib==3.0.1
python-dotenv==1.0.0
pydantic[email]==2.5.3
httpx==0.26.0
